attribution des adresses ip à partir de la plage d'adresses de la config

+++ src/com/dhcp/lease/Lease.java
@@ -13,7 +13,7 @@ public class Lease {
 	private long duration;
 	private boolean isAvailable = true;
 	
-	public Lease(InetAddress ipAddress, HardwareAddress hardwareAddress, int duration) {
+	public Lease(InetAddress ipAddress, HardwareAddress hardwareAddress, long duration) {
 		this.ipAddress = ipAddress;
 		this.currentHardwareAddress = hardwareAddress;
 		this.knownHardwareAddresses.add(hardwareAddress);
